+++ Lutra/Lutra.API/Requests/UpdateVerspakketRequest.cs
@@ -13,7 +13,7 @@ public sealed record UpdateVerspakketRequest(
     [Required] Guid SupermarktId,
     IReadOnlyList<VerspakketFotoRequest>? Fotos = null,
     IReadOnlyList<IngredientRequest>? Ingredienten = null,
-    VoedingswaardeRequest? Voedingswaarde = null,
+    IReadOnlyList<VoedingswaardeRequest>? Voedingswaarden = null,
     IReadOnlyList<Domain.Entities.Allergeen>? Allergenen = null);