+++ Lutra/Lutra.Application/Verspakketten/CreateVerspakket.Command.cs
@@ -13,6 +13,6 @@ public sealed partial class CreateVerspakket
         Beoordeling? Beoordeling,
         IReadOnlyList<VerspakketFoto>? Fotos = null,
         IReadOnlyList<Ingredient>? Ingredienten = null,
-        Voedingswaarde? Voedingswaarde = null,
+        IReadOnlyList<Voedingswaarde>? Voedingswaarden = null,
         IReadOnlyList<Domain.Entities.Allergeen>? Allergenen = null) : ICommand<Response>;
 }