+++ Lutra/Lutra.Domain/Entities/Verspakket.cs
@@ -7,6 +7,7 @@ public class Verspakket : BaseEntity
     private readonly List<Beoordeling> _beoordelingen = [];
     private readonly List<VerspakketFoto> _fotos = [];
     private readonly List<Ingredient> _ingredienten = [];
+    private readonly List<Voedingswaarde> _voedingswaarden = [];
     private readonly List<VerspakketAllergeen> _allergenen = [];
 
     [MaxLength(50)]
@@ -22,13 +23,13 @@ public class Verspakket : BaseEntity
 
     public virtual Supermarkt Supermarkt { get; set; } = null!;
 
-    public Voedingswaarde? Voedingswaarde { get; set; }
-
     public IReadOnlyCollection<Beoordeling> Beoordelingen => _beoordelingen.AsReadOnly();
 
     public IReadOnlyCollection<VerspakketFoto> Fotos => _fotos.AsReadOnly();
 
     public IReadOnlyCollection<Ingredient> Ingredienten => _ingredienten.AsReadOnly();
+
+    public IReadOnlyCollection<Voedingswaarde> Voedingswaarden => _voedingswaarden.AsReadOnly();
 
     public IReadOnlyCollection<VerspakketAllergeen> Allergenen => _allergenen.AsReadOnly();
 
@@ -45,6 +46,11 @@ public class Verspakket : BaseEntity
     public void AddIngredient(Ingredient ingredient)
     {
         _ingredienten.Add(ingredient);
+    }
+
+    public void AddVoedingswaarde(Voedingswaarde voedingswaarde)
+    {
+        _voedingswaarden.Add(voedingswaarde);
     }
 
     public void AddAllergeen(VerspakketAllergeen allergeen)