+++ Lutra/Lutra.Domain/Entities/VoedingswaardeBasis.cs
@@ -1,0 +1,7 @@
+namespace Lutra.Domain.Entities;
+
+public enum VoedingswaardeBasis
+{
+    Per100Gram,
+    PerPortie
+}